+++ src/index.ts
@@ -1,20 +1,14 @@
 export * as EmojiPicker from "./components/emoji-picker";
 export { defaultEmojiDataResolver } from "./data/emoji";
-export {
-  type CreateEmojiDataCacheOptions,
-  createEmojiDataCache,
-} from "./data/emoji-data-cache";
+export { createEmojiDataCache } from "./data/emoji-data-cache";
+export { getEmojiDetails } from "./data/get-emoji-details";
 export { useActiveEmoji, useSkinTone } from "./hooks";
 export type {
   Category,
   Emoji,
-  EmojibaseEmoji,
   EmojiData,
-  EmojiDataCache,
-  EmojiDataCategory,
-  EmojiDataEmoji,
   EmojiDataResolver,
-  EmojiDataResolverOptions,
+  EmojiDetails,
   EmojiPickerActiveEmojiProps,
   EmojiPickerEmptyProps,
   EmojiPickerListCategoryHeaderProps,